+++ src/components/Home/Home.css
@@ -1,31 +1,3 @@
-* {
-  margin: 0px;
-  padding: 0px;
-}
-
-.header__logo {
-  width: 80px;
-  background-color: #f7c6c6;
-  transform: rotate3d(0, 1, 1, -10deg) scale(1.1) translateY(-25px)
-    translateX(20px);
-}
-
-.header__logo-content {
-  font-size: 0.9em;
-  font-family: Courier New;
-  width: 30px;
-  text-align: center;
-  background-color: #8ecaf4;
-  color: #2c2f2a;
-  padding: 5px;
-  transform: rotate3d(0, 1, 1, 10deg) scale(1);
-}
-
-.header__upper-line {
-  background-color: #a9e8a9;
-  height: 37px;
-}
-
 .filters-block {
   padding: 20px;
   background-color: #a9eeea;
